Tasks now have pet_id and user_id; Pet and User lookup by name

+++ app/src/main/java/northeastern/is4300/pettasktracker/data/DatabaseHelper.java
@@ -3,6 +3,7 @@ package northeastern.is4300.pettasktracker.data;
 import android.content.Context;
 import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteOpenHelper;
+import android.util.Log;
 
 /**
  */
@@ -10,13 +11,10 @@ import android.database.sqlite.SQLiteOpenHelper;
 public class DatabaseHelper extends SQLiteOpenHelper {
 
     private static final String LOG = "DatabaseHelper";
-    private static final int DATABASE_VERSION = 1;
+    private static final int DATABASE_VERSION = 2;
     private static final String DATABASE_NAME = "petTaskTracker";
 
     private static final String KEY_ID = "id";
-    private static final String PET_ID = "pet_id";
-    private static final String USER_ID = "user_id";
-    private static final String TASK_ID = "user_id";
 
     private static final String CREATE_TABLE_PETS = "CREATE TABLE "
             + PetRepository.TABLE_PETS
@@ -35,7 +33,17 @@ public class DatabaseHelper extends SQLiteOpenHelper {
             + "(" + KEY_ID + " INTEGER PRIMARY KEY AUTOINCREMENT, "
             + TaskRepository.KEY_TASK_TYPE + " TEXT, "
             + TaskRepository.KEY_TASK_TIME + " TEXT, "
-            + TaskRepository.KEY_TASK_REPEAT + " TEXT)";
+            + TaskRepository.KEY_TASK_REPEAT + " TEXT, "
+            + TaskRepository.KEY_PET_ID + " INTEGER, "
+            + TaskRepository.KEY_USER_ID + " INTEGER)";
+
+    private static final String DATABASE_ALTER_TASK_TO_V2_1 = "ALTER TABLE "
+            + TaskRepository.TABLE_TASKS + " ADD COLUMN "
+            + TaskRepository.KEY_PET_ID + " INTEGER; ";
+
+    private static final String DATABASE_ALTER_TASK_TO_V2_2 = "ALTER TABLE "
+            + TaskRepository.TABLE_TASKS + " ADD COLUMN "
+            + TaskRepository.KEY_USER_ID + " INTEGER; ";
 
     public DatabaseHelper(Context context) {
         super(context, DATABASE_NAME, null, DATABASE_VERSION);
@@ -49,13 +57,16 @@ public class DatabaseHelper extends SQLiteOpenHelper {
     }
 
     @Override
-    public void onUpgrade(SQLiteDatabase db, int i, int i1) {
-        // on upgrade drop older tables
-        db.execSQL("DROP TABLE IF EXISTS " + PetRepository.TABLE_PETS);
-        db.execSQL("DROP TABLE IF EXISTS " + UserRepository.TABLE_USERS);
-        db.execSQL("DROP TABLE IF EXISTS " + TaskRepository.TABLE_TASKS);
+    public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
 
-        // create new tables
-        onCreate(db);
+        if (newVersion > oldVersion) {
+            Log.w(LOG,
+                    "Upgrading database from version " + oldVersion + " to "
+                    + newVersion + ", which will destroy all old data");
+            db.execSQL("DROP TABLE IF EXISTS " + TaskRepository.TABLE_TASKS);
+            onCreate(db);
+            db.execSQL(DATABASE_ALTER_TASK_TO_V2_1);
+            db.execSQL(DATABASE_ALTER_TASK_TO_V2_2);
+        }
     }
 }
